# specs/001-architect-client-portal/contracts/api.yaml
openapi: "3.1.0"
info:
  title: ArchiDrive API
  version: "1.0.0"
  description: Architect-Client Design Portal — REST + WebSocket API
servers:
  - url: http://localhost:8000
    description: Development
security:
  - bearerAuth: []
paths:
...
  /api/firms:
    post:
      tags: [Firms]
      summary: Create a firm (becomes firm admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "201":
          description: Firm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirmResponse"
    get:
      tags: [Firms]
      summary: List firms (firm admins see their firm)
      responses:
        "200":
          description: Firm list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FirmResponse"
  /api/firms/{firm_id}/members:
    get:
      tags: [Firms]
      summary: List firm members
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      tags: [Firms]
      summary: Add architect to firm (firm admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Member added }
        "404": { description: User not found }

  # ── Projects ──────────────────────────────────────────────
...
  /api/files/upload-url:
    post:
      tags: [Files]
      summary: Get presigned S3 upload URL (≤100MB files, architect only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, filename, content_type, file_size]
              properties:
                project_id: { type: integer }
                milestone_id: { type: integer, nullable: true }
                filename: { type: string }
                content_type: { type: string }
                file_size: { type: integer }
      responses:
        "200":
          description: Upload URL and file key
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, description: "Presigned PUT URL" }
                  key: { type: string, description: "S3 object key" }
                  file_id: { type: string, format: uuid }
  /api/files/multipart/initiate:
    post:
      tags: [Files]
      summary: Initiate S3 multipart upload (>100MB, architect only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, filename, content_type, file_size, part_size]
              properties:
                project_id: { type: integer }
                milestone_id: { type: integer, nullable: true }
                filename: { type: string }
                content_type: { type: string }
                file_size: { type: integer }
                part_size: { type: integer, minimum: 5242880 }
      responses:
        "200":
          description: Upload initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_id: { type: string }
                  key: { type: string }
                  file_id: { type: string, format: uuid }
  /api/files/multipart/{upload_id}/part-urls:
    post:
      tags: [Files]
      summary: Get presigned URLs for multipart upload parts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, part_numbers]
              properties:
                key: { type: string }
                part_numbers:
                  type: array
                  items: { type: integer }
                  minItems: 1
                  maxItems: 100
      responses:
        "200":
          description: Part URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: object
                    additionalProperties: { type: string }
  /api/files/multipart/{upload_id}/complete:
    post:
      tags: [Files]
      summary: Complete multipart upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, parts]
              properties:
                key: { type: string }
                parts:
                  type: array
                  items:
                    type: object
                    required: [PartNumber, ETag]
                    properties:
                      PartNumber: { type: integer }
                      ETag: { type: string }
      responses:
        "200": { description: Upload completed, thumbnail generation queued }
  /api/files/multipart/{upload_id}/abort:
    post:
      tags: [Files]
      summary: Abort incomplete multipart upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
      responses:
        "200": { description: Aborted }
...
  /api/files/{file_id}/thumbnail:
    get:
      tags: [Files]
      summary: Get thumbnail (redirect to presigned S3 URL)
      parameters:
        - in: query
          name: size
          schema: { type: string, enum: [small, medium], default: small }
      responses:
        "302": { description: Redirect to S3 URL }
        "404": { description: Thumbnail not available }
  /api/files/{file_id}/preview:
    get:
      tags: [Files]
      summary: Get 3D preview (redirect to presigned S3 URL for glTF/GLB)
      responses:
        "302": { description: Redirect to S3 URL }
        "404": { description: Preview not available }

  # ── Milestones ────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [architect, client] }
        firm_id: { type: integer, nullable: true }
        is_firm_admin: { type: boolean }
        is_verified: { type: boolean }
        created_at: { type: string, format: date-time }
    FirmResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        member_count: { type: integer }
        created_at: { type: string, format: date-time }
...